+++ tests/tests/test_color.py
@@ -1,27 +1,8 @@
-import pytest
+import copy
+import pickle
+from decimal import Decimal
 
 from spectrum.color import Color
-from spectrum.exceptions import InvalidOpacityError
-
-
-class TestColorComparison:
-    def setup_class(self):
-        self.color = Color("#D2CEB080")
-
-    def test_string(self):
-        assert self.color == "#D2CEB080"
-
-    def test_iterable(self):
-        assert self.color == [210, 206, 176, 128]
-
-    def test_color(self):
-        assert self.color == Color([210, 206, 176, 128])
-
-    def test_other_opacity(self):
-        assert self.color != Color([210, 206, 176])
-
-    def test_invalid_type(self):
-        assert self.color != 192
 
 
 class TestColor:
@@ -40,6 +21,9 @@ class TestColor:
     def test_alpha(self):
         assert self.color.alpha == 128
 
+    def test_opacity(self):
+        assert self.color.opacity == Decimal("0.5")
+
     def test_opaque(self):
         assert self.color.opaque is False
         assert Color("dadada").opaque is True
@@ -49,55 +33,75 @@ class TestColor:
         assert Color("dadada00").transparent is True
 
     def test_hex(self):
-        assert self.color.hex() == "#D2CEB0"
-
-    def test_hex_setter(self):
-        other_color = self.color.hex("ffff00")
-        assert other_color == [255, 255, 0, 128]
-        assert self.color is not other_color
+        assert self.color.hex == "#D2CEB0"
 
     def test_hexa(self):
-        assert self.color.hexa() == "#D2CEB080"
-
-    def test_hexa_setter(self):
-        other_color = self.color.hexa("aabbccff")
-        assert other_color == [170, 187, 204, 255]
-        assert self.color is not other_color
-
-    def test_opacity(self):
-        assert self.color.opacity() == 0.5
-
-    def test_opacity_setter(self):
-        other_color = self.color.opacity(0.1)
-        assert other_color == [210, 206, 176, 26]
-        assert self.color is not other_color
-
-    def test_set_invalid_opacity(self):
-        with pytest.raises(InvalidOpacityError):
-            self.color.opacity(2)
+        assert self.color.hexa == "#D2CEB080"
 
     def test_rgb(self):
-        assert self.color.rgb() == "rgb(210, 206, 176)"
+        assert self.color.rgb == "rgb(210, 206, 176)"
 
     def test_rgba(self):
-        assert self.color.rgba() == "rgba(210, 206, 176, 0.5)"
+        assert self.color.rgba == "rgba(210, 206, 176, 0.5)"
 
     def test_hsl(self):
-        assert self.color.hsl() == "hsl(53, 27.4%, 75.7%)"
+        assert self.color.hsl == "hsl(53, 27%, 76%)"
 
     def test_hsla(self):
-        assert self.color.hsla() == "hsla(53, 27.4%, 75.7%, 0.5)"
+        assert self.color.hsla == "hsla(53, 27%, 76%, 0.5)"
+
+    def test_as_tuple(self):
+        assert self.color.as_tuple() == (210, 206, 176, 128)
 
 
-class TestColorString:
-    def test_opaque(self):
+class TestSerialization:
+    def test_str_opaque(self):
         color = Color("#da60ec")
         assert str(color) == "#DA60EC"
 
-    def test_transparent(self):
+    def test_str_transparent(self):
         color = Color("#da60ecde")
         assert str(color) == "rgba(218, 96, 236, 0.87)"
 
-    def test_repr(self):
-        assert repr(Color("#da60ec")) == "Color('#DA60EC')"
-        assert repr(Color("#da60ecde")) == "Color('#DA60ECDE')"
+    def test_repr_opaque(self):
+        color = Color("#da60ec")
+        assert repr(color) == "Color('#DA60EC')"
+
+    def test_repr_transparent(self):
+        color = Color("#da60ecde")
+        assert repr(color) == "Color('#DA60ECDE')"
+
+    def test_pickle(self):
+        color = Color("rgba(218, 96, 236, 0.87)")
+        unpickled = pickle.loads(pickle.dumps(color))
+        assert color == unpickled
+
+
+class TestCopy:
+    def setup_class(self):
+        self.color = Color("#D2CEB080")
+
+    def test_copy(self):
+        clone = copy.copy(self.color)
+        clone._rgba[1] = 220
+        assert self.color._rgba[1] == 206
+
+
+class TestComparison:
+    def setup_class(self):
+        self.color = Color("#D2CEB080")
+
+    def test_string(self):
+        assert self.color == "#D2CEB080"
+
+    def test_iterable(self):
+        assert self.color == [210, 206, 176, 128]
+
+    def test_color(self):
+        assert self.color == Color([210, 206, 176, 128])
+
+    def test_other_opacity(self):
+        assert self.color != Color([210, 206, 176])
+
+    def test_invalid_type(self):
+        assert self.color != 192