+++ tests/tests/test_forms.py
@@ -11,18 +11,18 @@ class ColorForm(forms.Form):
 class TestColorField:
     def test_valid_initial(self):
         form = ColorForm({
-            'color': '#daBAcc80',
+            "color": "#daBAcc80",
         })
 
         assert form.is_valid() is True
 
-        color = form.cleaned_data['color']
+        color = form.cleaned_data["color"]
         assert type(color) is Color
 
     def test_invalid_color(self):
         form = ColorForm({
-            'color': '#AABBCCDDEE',
+            "color": "#AABBCCDDEE",
         })
 
         assert form.is_valid() is False
-        assert form.errors['color'].data[0].code == 'invalid'
+        assert form.errors["color"].data[0].code == "invalid"